requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy
pandas
sqlalchemy
Flask
gunicorn
json5
datetime
databases
asyncpg
azure-core
azure-eventhub
azure-identity
azure-schemaregistry
azure-schemaregistry-avroserializer
fastapi
okta-jwt
okta-jwt-verifier
httptools==0.1.*
httpx
swifter
geopy
psycopg2-binary

